Switch direction that lineOffset counts from
lineOffset counted the number of lines the bottom displayed line was
offset from the oldest line in the history. I felt this was
unintuitive, and it also required the offset to be changed every time
a new line was added. This commit will change the offset to count from
the latest line in the history; an offset of zero means that the last
printed line is displayed. An added benefit is that scrolling is more
intuitive: scrolling by a positive value moves up/back in the history.

The offset is still changed in scrollToBottom(). I'm not sure if this
function is needed now that scrolling to the bottom is so simple. If I
end up keeping it, I'll add a check (in scrollToBottom or setOffset)
to make sure the offset isn't set redudantly.